Got auth-server to work and made changesto types and moved the types file
auth-server works now and is usable after "npm i", "npm build" and "npm start" at localhost:3001/api + (/user or /auth). Had to add 2 types files so npm build would not build the complete file path to server. Current apidocs is not for this project and should not be refered to instead it should be recreated later. auth-serve still needs to be checked for bugs