Added register functionality to LoginRegisterPopup and fixed a bug in backend
Added route and functions for LoginRegisterPopup to register new user also made changes to some functions to fit the current implementation better. Login is not possible currently as the form would need to be changed or have another solution. Fixed a bug that would throw an error if user was not found as this was not intended